bluetooth: classic: Fix BR_SMP chan cannot found issue
Currently, `BT_L2CAP_CID_BR_SMP` is defined by
`BT_L2CAP_CHANNEL_DEFINE`. It causes the BR SMP fixed
l2cap channel is defined in LE fixed l2cap channel
list.
When BR ACL connection established, the fixed BR SMP
l2cap channel cannot be initialized.

When then BR ACL link encrypted, the smp_pairing
request cannot be preformed due to the smp_br channel
cannot be found.

And it causes a result that the LE LTK cannot be
derived from BR link key.

Use `BT_L2CAP_BR_CHANNEL_DEFINE` instead of
`BT_L2CAP_CHANNEL_DEFINE` to define BR SMP fixed
l2cap channel.
